Avoid using the include directive twice in the same RST file
If two questionnaire directives both use the include directive to
the same file in the questionnaire bodies, then new Sphinx 4.1 and
Docutils 0.17 versions produce warning "circular inclusion in
include directive" in the course RST build. The second
questionnaire does not compile correctly and does not work in A+.

If we don't use the include directive twice, then at least
the first call is successful.