Fix mathematical errors in the quiz contents
In addition, change the hints. The inverse hint (!a, when
the option is not selected) does not work in randomized questions
since the hint activates even when the corresponding option
is not part of the subset shown to the user. If the user can not
see the option, seeing hints triggered by that option is
confusing.